added thumbnail to model and thumnail functions to tasks and signals

+++ videoflix/content/signals.py
@@ -1,7 +1,7 @@
 import os
 import django_rq
 from .models import Video, Video480p, Video720p, Video1080p
-from .tasks import convert_480p, convert_720p, convert_1080p
+from .tasks import convert_480p, convert_720p, convert_1080p, create_thumbnail
 from django.dispatch import receiver
 from django.db.models.signals import post_save, post_delete
 
@@ -15,6 +15,7 @@ def video_post_save(sender, instance, created, **kwargs):
         queue.enqueue(convert_480p, instance.video_file.path)
         queue.enqueue(convert_720p, instance.video_file.path)
         queue.enqueue(convert_1080p, instance.video_file.path)
+        queue.enqueue(create_thumbnail, instance.video_file.path)
 
         video_480p_instance = Video480p(video=instance)
         video_480p_instance.video_file_480p = instance.video_file.path.replace(".mp4", "_480p.mp4")
@@ -28,12 +29,18 @@ def video_post_save(sender, instance, created, **kwargs):
         video_1080p_instance.video_file_1080p = instance.video_file.path.replace(".mp4", "_1080p.mp4")
         video_1080p_instance.save()
 
-@receiver(post_delete, sender=Video)
+        thumbnail_path = instance.video_file.path.replace(".mp4", "_thumbnail.jpg")
+        instance.thumbnail.name = thumbnail_path.replace("videos", "thumbnails")
+        instance.save()
+
 def auto_delete_file_on_delete(sender, instance, **kwargs):
     """
-    Deletes file from filesystem
-    when corresponding `Video` object is deleted.
+    Deletes the video file, converted files, and thumbnail when the Video object is deleted.
     """
-    if instance.video_file:
-        if os.path.isfile(instance.video_file.path):
-            os.remove(instance.video_file.path)
+    if instance.video_file and os.path.isfile(instance.video_file.path):
+        os.remove(instance.video_file.path)
+
+    for resolution in ['_480p.mp4', '_720p.mp4', '_1080p.mp4', '_thumbnail.jpg']:
+        file_path = instance.video_file.path.replace(".mp4", resolution)
+        if os.path.isfile(file_path):
+            os.remove(file_path)
